+++ index.js
@@ -16,6 +16,7 @@
 │   limitations under the License.                                            │
 \*───────────────────────────────────────────────────────────────────────────*/
 'use strict';
+var crypto  = require('crypto');
 
 
 /**
@@ -23,11 +24,16 @@
  * @param {Object} options The configuration object.
  */
 var lusca = module.exports = function (options) {
-    var headers = [];
+    var headers = [], nonce;
 
     if (options) {
         Object.keys(lusca).forEach(function (key) {
             var config = options[key];
+            if (key === "csp" && options[key] && (options[key]['styleNonce'] || options[key]['scriptNonce'])) {
+                nonce = true;
+            }
+
+            //if (key === 'csp' && lusca[key](config) && lusca[key](config)[styleNonce]) 
 
             if (config) {
                 headers.push(lusca[key](config));
@@ -38,6 +44,12 @@ var lusca = module.exports = function (options) {
     return function lusca(req, res, next) {
         var chain = next;
 
+        if (nonce) {
+            Object.defineProperty(res.locals, 'nonce', {
+                value: crypto.pseudoRandomBytes(36).toString('base64'),
+                enumerable: true
+            });
+        }
         headers.forEach(function (header) {
             chain = (function (next) {
                 return function (err) {
@@ -46,6 +58,7 @@ var lusca = module.exports = function (options) {
                         return;
                     }
                     header(req, res, next);
+                    return;
                 };
             }(chain));
         });