+++ test/mocks/config/nonce.js
@@ -1,0 +1,11 @@
+'use strict';
+
+
+module.exports = {
+	reportOnly: false,
+	scriptNonce: true,
+	policy: {
+		"default-src": "*",
+		"script-src": "'unsafe-inline"
+	}
+};